Split up grunt tasks (fixes #32)
