+++ config/devserver.js
@@ -1,0 +1,8 @@
+module.exports = {
+  options: {
+    base: 'public',
+    type: 'http',
+    port: 4000
+  },
+  server: {}
+};
