+++ config/gh-pages.js
@@ -1,0 +1,8 @@
+module.exports = {
+  options: {
+    base: 'styleguide/build',
+    message: 'Updating docs',
+    repo: 'git@github.com:mozilla/payments-ui.git'
+  },
+  src: ['**']
+};
