+++ config/jshint.js
@@ -1,0 +1,6 @@
+module.exports = {
+  options: {
+    jshintrc: true
+  },
+  all: ['public/js/*.js', 'test/**/*.js']
+};
