+++ config/react.js
@@ -1,0 +1,11 @@
+module.exports = {
+  jsx: {
+    files: [{
+      expand: true,
+      cwd: 'jsx',
+      src: ['**/*.jsx'],
+      dest: 'public/js/components/',
+      ext: '.js'
+    }]
+  }
+};
