+++ config/sass.js
@@ -1,0 +1,10 @@
+module.exports = {
+  options: {
+    sourceMap: true
+  },
+  dist: {
+    files: {
+      'public/css/main.css': 'public/scss/main.scss'
+    }
+  }
+};
