+++ config/watch.js
@@ -1,0 +1,16 @@
+module.exports = {
+  sass: {
+    files: 'public/scss/**/*.scss',
+    tasks: ['sass'],
+    options: {
+      debounceDelay: 250
+    }
+  },
+  jsx: {
+    files: 'jsx/**/*.jsx',
+    tasks: ['react'],
+    options: {
+      debounceDelay: 250
+    }
+  }
+};
